datavex-main/info_gather/datavex-srivatsa/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic-settings==2.5.2
python-dotenv==1.0.1
lxml==5.3.0
duckduckgo-search==6.3.7
